+++ relay-server/src/processing/transactions/mod.rs
@@ -145,82 +145,21 @@ impl Processor for TransactionProcessor {
         work: Managed<Self::UnitOfWork>,
         mut ctx: super::Context<'_>,
     ) -> Result<super::Output<Self::Output>, Rejected<Self::Error>> {
+        let scoping = work.scoping();
         let project_id = work.scoping().project_id;
-        let transaction = work.transaction.as_ref();
-        let flags = Flags {
-            metrics_extracted: transaction.is_some_and(|i| i.metrics_extracted()),
-            spans_extracted: transaction.is_some_and(|i| i.spans_extracted()),
-            fully_normalized: work.headers.meta().request_trust().is_trusted()
-                && work
-                    .transaction
-                    .as_ref()
-                    .is_some_and(Item::fully_normalized),
-        };
         let mut metrics = Metrics::default();
         let mut extracted_metrics = ProcessingExtractedMetrics::new();
 
-        let mut work = work.try_map(|w, _| {
-            let SerializedTransaction {
-                headers,
-                transaction: transaction_item,
-                attachments,
-                profile,
-            } = w;
-            let mut transaction = Annotated::empty();
-            if let Some(transaction_item) = transaction_item.as_ref() {
-                transaction = metric!(timer(RelayTimers::EventProcessingDeserialize), {
-                    Annotated::<Event>::from_json_bytes(&transaction_item.payload())
-                })?;
-                if let Some(event) = transaction.value_mut() {
-                    event.ty = EventType::Transaction.into();
-                }
-            }
-            Ok::<_, Error>(ExpandedTransaction {
-                headers,
-                transaction: Transaction(transaction),
-                flags,
-                attachments,
-                profile,
-                extracted_spans: vec![],
-            })
-        })?;
+        let mut work = work.try_map(|w, _| process::expand(w))?;
 
-        let mut profile_id = None;
         let mut run_dynamic_sampling = false;
         work.try_modify(|work, record_keeper| {
-            if let Some(profile_item) = work.profile.as_mut() {
-                let feature = Feature::Profiling;
-                if should_filter(ctx.config, ctx.project_info, feature) {
-                    record_keeper.reject_err(
-                        Outcome::Invalid(DiscardReason::FeatureDisabled(feature)),
-                        work.profile.take(),
-                    );
-                } else if work.transaction.0.value().is_none() && profile_item.sampled() {
-                    // A profile with `sampled=true` should never be without a transaction
-                    record_keeper.reject_err(
-                        Outcome::Invalid(DiscardReason::Profiling("missing_transaction")),
-                        work.profile.take(),
-                    );
-                } else {
-                    match relay_profiling::parse_metadata(&profile_item.payload(), project_id) {
-                        Ok(id) => {
-                            profile_id = Some(id);
-                        }
-                        Err(err) => {
-                            record_keeper.reject_err(
-                                Outcome::Invalid(DiscardReason::Profiling(
-                                    relay_profiling::discard_reason(err),
-                                )),
-                                work.profile.take(),
-                            );
-                        }
-                    }
-                }
-            }
+
+            let profile_id = profile::filter(work, record_keeper, &ctx, project_id);
 
             let event = &mut work.transaction.0;
             profile::transfer_id(event, profile_id);
-            // TODO: profile::remove_context_if_rate_limited(&mut event, managed_envelope, ctx);
+            profile::remove_context_if_rate_limited(event, scoping, ctx);
 
             utils::dsc::validate_and_set_dsc(&mut work.headers, event, &mut ctx);
 
@@ -283,72 +222,19 @@ impl Processor for TransactionProcessor {
         let server_sample_rate = sampling_result.sample_rate();
 
         if let Some(outcome) = sampling_result.into_dropped_outcome() {
-            // Process profiles before dropping the transaction, if necessary.
-            // Before metric extraction to make sure the profile count is reflected correctly.
-
-            work.try_modify(|work, r| {
-                if let Some(profile) = work.profile.as_mut() {
-                    profile.set_sampled(false);
-                    let result = profile::process(
-                        profile,
-                        work.headers.meta().client_addr(),
-                        work.transaction.0.value(),
-                        &ctx,
-                    );
-                    if let Err(outcome) = result {
-                        r.reject_err(outcome, work.profile.take());
-                    }
-                }
-                // Extract metrics here, we're about to drop the event/transaction.
-                work.flags.metrics_extracted = utils::transaction::extract_metrics(
-                    &mut work.transaction.0,
-                    &mut extracted_metrics,
-                    ExtractMetricsContext {
-                        dsc: work.headers.dsc(),
-                        project_id,
-                        ctx: &ctx,
-                        sampling_decision: SamplingDecision::Drop,
-                        metrics_extracted: EventMetricsExtracted(work.flags.metrics_extracted),
-                        spans_extracted: work.flags.spans_extracted,
-                    },
-                )?
-                .0;
-                Ok::<_, Error>(())
-            })?;
-
-            // .map_err(Error::ProcessingFailed)
-            // .reject(&work)?;
-
-            let (work, profile) = work.split_once(|mut work| {
-                let profile = work.profile.take();
-                (work, profile)
-            });
-
-            // reject everything but the profile:
-            // FIXME: track non-extracted spans as well.
-            // -> Type TransactionWithEmbeddedSpans
-            let _ = work.reject_err(outcome);
-
-            // If we have a profile left, we need to make sure there is quota for this profile.
-            let profile = profile.transpose();
-            if let Some(profile) = profile {
-                let mut profile = profile.map(|p, _| Profile(p));
-                self.limiter.enforce_quotas(&mut profile, ctx).await?;
-            }
-
-            let metrics = work.wrap(extracted_metrics.into_inner());
-            return Ok(super::Output {
-                main: Some(TransactionOutput(work)),
-                metrics: Some(metrics),
-            });
+            return process::drop_after_sampling(
+                work,
+                extracted_metrics,
+                outcome,
+                ctx,
+                &self.limiter,
+            )
+            .await;
         }
-
-        // let _post_ds = cogs.start_category("post_ds"); // FIXME
 
         // Need to scrub the transaction before extracting spans.
         //
         // Unconditionally scrub to make sure PII is removed as early as possible.
-
         work.try_modify(|work, _| {
             utils::event::scrub(&mut work.transaction.0, ctx.project_info)?;
             utils::attachments::scrub(work.attachments.iter_mut(), ctx.project_info);
@@ -623,7 +509,7 @@ impl RateLimited for Managed<ExpandedTransaction> {
     }
 }
 
-#[derive(Debug)]
+#[derive(Debug, Default)]
 struct Flags {
     metrics_extracted: bool,
     spans_extracted: bool,
