+++ relay-server/src/processing/transactions/profile.rs
@@ -1,19 +1,23 @@
 //! Profiles related processor code.
+use relay_base_schema::project::ProjectId;
 use relay_dynamic_config::{Feature, GlobalConfig};
+use relay_quotas::{DataCategory, Scoping};
 use std::net::IpAddr;
 
 use relay_config::Config;
 use relay_event_schema::protocol::{Contexts, Event, ProfileContext};
 use relay_filter::ProjectFiltersConfig;
-use relay_profiling::ProfileId;
-use relay_protocol::Annotated;
+use relay_profiling::{ProfileId, ProfileType};
+use relay_protocol::{Annotated, Empty};
 #[cfg(feature = "processing")]
 use relay_protocol::{Getter, Remark, RemarkType};
 
 use crate::envelope::{ContentType, Item, ItemType};
-use crate::managed::{Counted, Managed};
+use crate::managed::{Counted, Managed, RecordKeeper};
+use crate::processing::transactions::ExpandedTransaction;
 use crate::processing::{Context, CountRateLimited};
 use crate::services::outcome::{DiscardReason, Outcome};
+use crate::utils::should_filter;
 
 pub struct Profile(pub Item);
 
@@ -25,6 +29,49 @@ impl Counted for Profile {
 
 impl CountRateLimited for Managed<Profile> {
     type Error = super::Error;
+}
+
+/// Filters out invalid profiles.
+///
+/// Returns the profile id of the single remaining profile, if there is one.
+pub fn filter(
+    work: &mut ExpandedTransaction,
+    record_keeper: &mut RecordKeeper,
+    ctx: &Context,
+    project_id: ProjectId,
+) -> Option<ProfileId> {
+    let Some(profile_item) = work.profile.as_ref() else {
+        return None;
+    };
+    let feature = Feature::Profiling;
+    let mut profile_id = None;
+    if should_filter(ctx.config, ctx.project_info, feature) {
+        record_keeper.reject_err(
+            Outcome::Invalid(DiscardReason::FeatureDisabled(feature)),
+            work.profile.take(),
+        );
+    } else if work.transaction.0.value().is_none() && profile_item.sampled() {
+        // A profile with `sampled=true` should never be without a transaction
+        record_keeper.reject_err(
+            Outcome::Invalid(DiscardReason::Profiling("missing_transaction")),
+            work.profile.take(),
+        );
+    } else {
+        match relay_profiling::parse_metadata(&profile_item.payload(), project_id) {
+            Ok(id) => {
+                profile_id = Some(id);
+            }
+            Err(err) => {
+                record_keeper.reject_err(
+                    Outcome::Invalid(DiscardReason::Profiling(relay_profiling::discard_reason(
+                        err,
+                    ))),
+                    work.profile.take(),
+                );
+            }
+        }
+    }
+    profile_id
 }
 
 /// Transfers the profile ID from the profile item to the transaction item.
@@ -52,6 +99,57 @@ pub fn transfer_id(event: &mut Annotated<Event>, profile_id: Option<ProfileId>) 
                 profile_context.profile_id = Annotated::empty();
             }
         }
+    }
+}
+
+/// Removes the profile context from the transaction item if there is an active rate limit.
+///
+/// With continuous profiling profile chunks are ingested separately to transactions,
+/// in the case where these profiles are rate limited the link on the associated transaction(s)
+/// should also be removed.
+///
+/// See also: <https://github.com/getsentry/relay/issues/5071>.
+pub fn remove_context_if_rate_limited(
+    event: &mut Annotated<Event>,
+    scoping: Scoping,
+    ctx: Context<'_>,
+) {
+    let Some(event) = event.value_mut() else {
+        return;
+    };
+
+    // There is always only either a transaction profile or a continuous profile, never both.
+    //
+    // If the `profiler_id` is set on the context, it is for a continuous profile, the case we want
+    // to handle here.
+    // If it is empty -> do nothing.
+    let profile_ctx = event.context::<ProfileContext>();
+    if profile_ctx.is_none_or(|pctx| pctx.profiler_id.is_empty()) {
+        return;
+    }
+
+    // Continuous profiling has two separate categories based on the platform, infer the correct
+    // category to check for rate limits.
+    let categories = match event.platform.as_str().map(ProfileType::from_platform) {
+        Some(ProfileType::Ui) => &[
+            scoping.item(DataCategory::ProfileChunkUi),
+            scoping.item(DataCategory::ProfileDurationUi),
+        ],
+        Some(ProfileType::Backend) => &[
+            scoping.item(DataCategory::ProfileChunk),
+            scoping.item(DataCategory::ProfileDuration),
+        ],
+        _ => return,
+    };
+
+    // This is a 'best effort' approach, which is why it is enough to check against cached rate
+    // limits here.
+    let is_limited = ctx
+        .rate_limits
+        .is_any_limited_with_quotas(ctx.project_info.get_quotas(), categories);
+
+    if is_limited && let Some(contexts) = event.contexts.value_mut() {
+        let _ = contexts.remove::<ProfileContext>();
     }
 }
