+++ relay-server/src/services/processor.rs
@@ -1459,7 +1459,11 @@ impl EnvelopeProcessorService {
             ctx.project_info,
         );
         processing::transactions::profile::transfer_id(&mut event, profile_id);
-        profile::remove_context_if_rate_limited(&mut event, managed_envelope, ctx);
+        processing::transactions::profile::remove_context_if_rate_limited(
+            &mut event,
+            managed_envelope.scoping(),
+            ctx,
+        );
 
         ctx.sampling_project_info = dynamic_sampling::validate_and_set_dsc(
             managed_envelope,
